+++ packages/storefront/src/lib/composables/use-component-variant.ts
@@ -1,0 +1,13 @@
+import { computed } from 'vue';
+
+export default (props: Record<string, any>) => {
+  return computed(() => {
+    let variantName = '';
+    Object.keys(props).forEach((prop) => {
+      if (props[prop] === true) {
+        variantName += ` ${prop.replace(/^(is|has)/, '')}`;
+      }
+    });
+    return variantName.slice(1);
+  });
+};